externalise ref debounce
