+++ app/lib/common/extensions/ref_debounce.dart
@@ -1,0 +1,22 @@
+import 'dart:async';
+
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+/// An extension on [Ref] with helpful methods to add a debounce.
+extension RefDebounceExtension on Ref {
+  /// Delays an execution by a bit such that if a dependency changes multiple
+  /// time rapidly, the rest of the code is only run once.
+  Future<void> debounce(Duration duration) {
+    final completer = Completer<void>();
+    final timer = Timer(duration, () {
+      if (!completer.isCompleted) completer.complete();
+    });
+    onDispose(() {
+      timer.cancel();
+      if (!completer.isCompleted) {
+        completer.completeError(StateError('Cancelled'));
+      }
+    });
+    return completer.future;
+  }
+}
