+++ app/lib/common/utils/utils.dart
@@ -4,7 +4,6 @@ import 'dart:ui';
 
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:jiffy/jiffy.dart';
 import 'package:logging/logging.dart';
@@ -33,25 +32,6 @@ bool isValidUrl(String url) {
     caseSensitive: false,
   );
   return urlPattern.hasMatch(url);
-}
-
-/// An extension on [Ref] with helpful methods to add a debounce.
-extension RefDebounceExtension on Ref {
-  /// Delays an execution by a bit such that if a dependency changes multiple
-  /// time rapidly, the rest of the code is only run once.
-  Future<void> debounce(Duration duration) {
-    final completer = Completer<void>();
-    final timer = Timer(duration, () {
-      if (!completer.isCompleted) completer.complete();
-    });
-    onDispose(() {
-      timer.cancel();
-      if (!completer.isCompleted) {
-        completer.completeError(StateError('Cancelled'));
-      }
-    });
-    return completer.future;
-  }
 }
 
 String jiffyTime(BuildContext context, int timeInterval) {
